+++ plugins/code-enhancers-and-utils/all-projects/main/kotlin/com/javiersc/gradle/plugins/all/projects/install/pre/commit/InstallAssemblePreCommitTask.kt
@@ -1,19 +1,24 @@
 package com.javiersc.gradle.plugins.all.projects.install.pre.commit
 
-import com.javiersc.gradle.plugins.all.projects.allProjectsExtension
 import org.gradle.api.Project
+import org.gradle.api.provider.Property
+import org.gradle.api.tasks.CacheableTask
 import org.gradle.api.tasks.Input
 import org.gradle.api.tasks.TaskAction
+import org.gradle.kotlin.dsl.register
 
-open class InstallAssemblePreCommitTask : InstallPreCommitTask() {
+@CacheableTask
+abstract class InstallAssemblePreCommitTask : InstallPreCommitTask() {
 
     @get:Input
     override val preCommitName: String
         get() = "assemble"
 
+    @get:Input abstract val assemble: Property<Boolean>
+
     @TaskAction
     override fun install() {
-        if (project.allProjectsExtension.install.get().preCommit.get().assemble.get()) {
+        if (assemble.get()) {
             createInstallPreCommitGradleTask()
         }
     }
@@ -21,12 +26,15 @@ open class InstallAssemblePreCommitTask : InstallPreCommitTask() {
     companion object {
         const val name: String = "installAssemblePreCommit"
 
-        internal fun register(project: Project) {
+        internal fun register(
+            project: Project,
+            configure: InstallAssemblePreCommitTask.() -> Unit
+        ) {
             val installAllTestsPreCommitTask =
-                project.tasks.register(name, InstallAssemblePreCommitTask::class.java) { task ->
-                    task.group = taskGroup
-
-                    task.finalizedBy(WriteFilePreCommitTask.getTask(project))
+                project.tasks.register<InstallAssemblePreCommitTask>(name) {
+                    group = taskGroup
+                    configure(this)
+                    finalizedBy(WriteFilePreCommitTask.getTask(project))
                 }
             getInstallPreCommitTask(project).get().dependsOn(installAllTestsPreCommitTask)
         }