+++ plugins/code-enhancers-and-utils/all-projects/main/kotlin/com/javiersc/gradle/plugins/all/projects/install/pre/commit/PreCommitOptions.kt
@@ -5,7 +5,7 @@ import org.gradle.api.model.ObjectFactory
 import org.gradle.api.provider.Property
 import org.gradle.kotlin.dsl.property
 
-open class PreCommitOptions @Inject constructor(objects: ObjectFactory) {
+abstract class PreCommitOptions @Inject constructor(objects: ObjectFactory) {
 
     val allTests: Property<Boolean> = objects.property<Boolean>().convention(true)
     val apiCheck: Property<Boolean> = objects.property<Boolean>().convention(true)