+++ plugins/code-enhancers-and-utils/all-projects/main/kotlin/com/javiersc/gradle/plugins/all/projects/install/pre/commit/WriteFilePreCommitTask.kt
@@ -3,22 +3,36 @@ package com.javiersc.gradle.plugins.all.projects.install.pre.commit
 import java.io.File
 import org.gradle.api.DefaultTask
 import org.gradle.api.Project
-import org.gradle.api.tasks.InputDirectory
+import org.gradle.api.tasks.CacheableTask
+import org.gradle.api.tasks.InputFile
+import org.gradle.api.tasks.Internal
 import org.gradle.api.tasks.OutputFile
+import org.gradle.api.tasks.PathSensitive
+import org.gradle.api.tasks.PathSensitivity
 import org.gradle.api.tasks.TaskAction
 import org.gradle.api.tasks.TaskProvider
+import org.gradle.kotlin.dsl.register
 
-open class WriteFilePreCommitTask : DefaultTask() {
+@CacheableTask
+abstract class WriteFilePreCommitTask : DefaultTask() {
 
-    @get:InputDirectory
-    val installPreCommitsDir: File =
-        project.file("${project.rootProject.buildDir}/install/pre-commits/")
+    @get:Internal
+    val installPreCommitsDir: File
+        get() =
+            project.file("${project.rootProject.buildDir}/install/pre-commits/").apply { mkdirs() }
 
-    @get:OutputFile val preCommitFile: File = project.preCommitFile
+    @get:InputFile
+    @get:PathSensitive(PathSensitivity.ABSOLUTE)
+    val inputPreCommitFile: File
+        get() = project.preCommitFile
+
+    @get:OutputFile
+    val outputPreCommitFile: File
+        get() = project.preCommitFile
 
     @TaskAction
     fun writeFile() {
-        val currentPreCommitText = preCommitFile.apply { createNewFile() }.readText()
+        val currentPreCommitText = inputPreCommitFile.apply { createNewFile() }.readText()
 
         val preCommitText: String =
             installPreCommitsDir
@@ -28,15 +42,15 @@ open class WriteFilePreCommitTask : DefaultTask() {
                     if (!currentPreCommitText.contains(text)) accumulative + text else accumulative
                 }
                 .orEmpty()
-        preCommitFile.writeText(currentPreCommitText + preCommitText)
+        outputPreCommitFile.writeText(currentPreCommitText + preCommitText)
     }
 
     companion object {
         const val name: String = "writeFilePreCommit"
 
         internal fun register(project: Project) {
-            project.tasks.register(name, WriteFilePreCommitTask::class.java) { task ->
-                task.group = InstallPreCommitTask.taskGroup
+            project.tasks.register<WriteFilePreCommitTask>(name) {
+                group = InstallPreCommitTask.taskGroup
             }
         }